Reorganized UiElement tests
